Camera movement implemented!
Interactive rendering implementation started.

+++ RayTracer/src/SpaceConversions.cpp
@@ -1,13 +1,13 @@
 #include "SpaceConversions.h"
 
 FVector2 world2RasterSpace( const FVector2& w_coords, const ImagePlane& imgPlane ) {
-    return { w_coords.x - (imgPlane.resolution.x / 2),
-             w_coords.y - (imgPlane.resolution.y / 2) };
+    return { w_coords.x - (imgPlane.resolution.x / 2.f),
+             w_coords.y - (imgPlane.resolution.y / 2.f) };
 }
 
 FVector2 raster2WorldSpace( const FVector2& w_coords, const ImagePlane& imgPlane ) {
-    return { imgPlane.resolution.x / 2 + abs( w_coords.x ),
-             imgPlane.resolution.y / 2 + abs( w_coords.y ) };
+    return { imgPlane.resolution.x / 2.f + abs( w_coords.x ),
+             imgPlane.resolution.y / 2.f + abs( w_coords.y ) };
 }
 
 FVector2 raster2NDC( const FVector2& coords, const ImagePlane& imgPlane ) {
@@ -15,29 +15,29 @@ FVector2 raster2NDC( const FVector2& coords, const ImagePlane& imgPlane ) {
 }
 
 FVector2 ray2NDC( const int x, const int y, const ImagePlane& imgPlane ) {
-    return { (x + 0.5) / imgPlane.resolution.x, (y + 0.5) / imgPlane.resolution.y };
+    return { (x + 0.5f) / imgPlane.resolution.x, (y + 0.5f) / imgPlane.resolution.y };
 }
 
 FVector2 NDC2ScreenSpace( const FVector2& coords ) {
     // Inverts Y
-    return { (2.0 * coords.x) - 1.0, 1.0 - (2.0 * coords.y) };
+    return { (2.f * coords.x) - 1.f, 1.f - (2.f * coords.y) };
     // No-invert version
-     //return { (2.0 * coords.x) - 1.0, (2.0 * coords.y) -1.0 };
+     //return { (2.f * coords.x) - 1.f, (2.f * coords.y) -1.f };
 }
 
 FVector2 ScreenSpace2NDC( const FVector2& coords ) {
     // Inverts Y
-    return { (coords.x + 1.0) / 2, (1.0 - coords.y) / 2 };
+    return { (coords.x + 1.f) / 2.f, (1.f - coords.y) / 2.f };
     // No-invert version
-    //return { (coords.x + 1.0) / 2, (coords.y + 1.0) / 2 };
+    //return { (coords.x + 1.f) / 2.f, (coords.y + 1.f) / 2.f };
 }
 
-double ScreenSpace2NDC( double val ) {
-    return (val + 1.0) / 2;
+float ScreenSpace2NDC( float val ) {
+    return (val + 1.f) / 2.f;
 }
 
-double ScreenSpace2NDCInvert( double val ) {
-    return (1.0 - val) / 2;
+float ScreenSpace2NDCInvert( float val ) {
+    return (1.f - val) / 2.f;
 }
 
 FVector2 getFixedAspectRatio( const FVector2& vec, const ImagePlane& imgPlane ) {
